requirements.txt
Adafruit-Blinka==8.16.1
adafruit-circuitpython-busdevice==5.2.4
adafruit-circuitpython-dht==3.7.8
adafruit-circuitpython-requests==1.13.1
adafruit-circuitpython-typing==1.9.0
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==3.43.0
Adafruit-PureIO==1.1.10
asgiref==3.6.0
board==1.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
idna==3.4
numpy==1.24.2
pandas==2.0.0
psycopg2==2.9.5
pyftdi==0.54.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3
pyusb==1.2.1
requests==2.28.2
RPi.GPIO==0.7.1
rpi-ws281x==4.3.4
setuptools==44.1.1
six==1.16.0
sqlparse==0.4.3
sysv-ipc==1.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
